Refactored utils into schema2CedarBase
